# Conflicts:
#	models/Document/Tag/Video.php

+++ models/DataObject/Concrete/Dao.php
@@ -143,6 +143,10 @@ class Dao extends Model\DataObject\AbstractObject\Dao
      */
     public function getData()
     {
+        if (empty($this->model->getClass())) {
+            return;
+        }
+
         $data = $this->db->fetchRow('SELECT * FROM object_store_' . $this->model->getClassId() . ' WHERE oo_id = ?', $this->model->getId());
 
         $fieldDefinitions = $this->model->getClass()->getFieldDefinitions(['object' => $this->model]);
@@ -154,6 +158,8 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                         'context' => [
                             'object' => $this->model,
                         ],
+                        'owner' => $this->model,
+                        'fieldname' => $key,
                     ];
                     $value = $value->load($this->model, $params);
                     if ($value === 0 || !empty($value)) {
@@ -170,7 +176,9 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                     }
                     $this->model->setValue($key, $value->getDataFromResource($multidata));
                 } else {
-                    $this->model->setValue($key, $value->getDataFromResource($data[$key], $this->model));
+                    $this->model->setValue($key, $value->getDataFromResource($data[$key], $this->model,
+                            ['owner' => $this->model,
+                                'fieldname' => $key, ]));
                 }
             }
         }
@@ -236,7 +244,11 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                     'isUpdate' => $isUpdate,
                     'context' => [
                         'containerType' => 'object',
-                    ], ];
+                    ],
+                    'owner' => $this->model,
+                    'fieldname' => $key,
+                ]
+                ;
                 if ($this->model instanceof Model\Element\DirtyIndicatorInterface) {
                     $saveParams['newParent'] = $this->model->isFieldDirty('o_parentId');
                 }
@@ -249,6 +261,7 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                         [
                             'isUpdate' => $isUpdate,
                             'owner' => $this->model,
+                            'fieldname' => $key,
                         ]);
                     if (is_array($insertDataArray)) {
                         $data = array_merge($data, $insertDataArray);
@@ -258,6 +271,7 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                         [
                             'isUpdate' => $isUpdate,
                             'owner' => $this->model,
+                            'fieldname' => $key,
                         ]);
                     $data[$key] = $insertData;
                 }
@@ -296,6 +310,7 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                         [
                             'isUpdate' => $isUpdate,
                             'owner' => $this->model,
+                            'fieldname' => $key,
                         ]);
                     $isEmpty = $fd->isEmpty($fieldValue);
 
@@ -404,9 +419,11 @@ class Dao extends Model\DataObject\AbstractObject\Dao
         $this->db->delete('object_relations_' . $this->model->getClassId(), ['src_id' => $this->model->getId()]);
 
         // delete fields which have their own delete algorithm
-        foreach ($this->model->getClass()->getFieldDefinitions() as $fd) {
-            if ($fd instanceof CustomResourcePersistingInterface) {
-                $fd->delete($this->model);
+        if ($this->model->getClass()) {
+            foreach ($this->model->getClass()->getFieldDefinitions() as $fd) {
+                if ($fd instanceof CustomResourcePersistingInterface) {
+                    $fd->delete($this->model);
+                }
             }
         }
